validation

+++ Controllers/Bank.controller.js
@@ -1,6 +1,5 @@
 const fs = require("fs");
 const bankJson = require("./../Json/Bank.json");
-
 const { users } = bankJson;
 
 // Show details of all users
@@ -12,21 +11,20 @@ const getAllUsers = (req, res) => {
 const getUserById = (req, res) => {
     const { id } = req.params;
     const user = users.find((user) => user.id === parseInt(id));
-    console.log(user);
     if (user) return res.status(200).json({ user });
     return res.status(200).json({ error: 'Sorry, user not found' })
 };
 
 // Add user
 const addUser = (req, res) => {
-    // const id = req.body.id;
     const { id } = req.body;
-    // console.log(String(req.body.id).length);
-    // const { cash = 0, credit = 0 } = req.body;
-    users.push({ id, cash: 0, credit: 0 });
-    const jsonData = JSON.stringify(bankJson);
-    fs.writeFileSync("Json/Bank.json", jsonData);
-    return res.status(200).json({ success: "user add to db" });
+    if (!checkifUserExist(id)) {
+        users.push({ id, cash: 0, credit: 0 });
+        const jsonData = JSON.stringify(bankJson);
+        fs.writeFileSync("Json/Bank.json", jsonData);
+        return res.status(200).json({ success: "user add to db" });
+    }
+    return res.status(200).json({ error: "this id is alreade exist in db" });
 };
 
 //Depositing
@@ -34,18 +32,16 @@ const deposit = (req, res) => {
     const { id } = req.params;
     const { amount } = req.body;
     const user = users.find((user) => user.id === parseInt(id));
-    console.log(user);
-    if (user) {
-        if (amount > 0) {
+    if (checkifUserExist(id)) {
+        if (!checkAmount(amount).error) {
             user.cash += amount;
             const jsonData = JSON.stringify(bankJson);
             fs.writeFileSync("Json/Bank.json", jsonData);
             return res.status(200).json("Success")
         }
-        return res.status(200).json({ error: "Check your amount again- cannot be a negative number or proide value" })
-
+        return res.json({ error: checkAmount(amount).error })
     }
-    return res.status(200).json({ error: "Sorry, user not found " })
+    return res.status(200).json({ error: "Sorry, user not found" })
 };
 
 //Update credit
@@ -53,16 +49,14 @@ const updateCredit = (req, res) => {
     const { id } = req.params;
     const { amount } = req.body;
     const user = users.find((user) => user.id === parseInt(id));
-    console.log(user);
-    if (user) {
-        if (amount > 0) {
+    if (checkifUserExist(id)) {
+        if (!checkAmount(amount).error) {
             user.credit += amount;
             const jsonData = JSON.stringify(bankJson);
             fs.writeFileSync("Json/Bank.json", jsonData);
             return res.status(200).json("Success")
         }
-        return res.status(200).json({ error: "Check your amount again- cannot be a negative number or proide value" })
-
+        return res.json({ error: checkAmount(amount).error })
     }
     return res.status(200).json({ error: "Sorry, user not found " })
 };
@@ -73,9 +67,9 @@ const withdrawMoney = (req, res) => {
     const { amount } = req.body;
     const user = users.find((user) => user.id === parseInt(id));
     console.log(user);
-    if (user) {
-        if (amount > 0) {
-            if (user.cash + user.credit < amount) {
+    if (checkifUserExist(id)) {
+        if (!checkAmount(amount).error) {
+            if (checkWithdraw(id, amount)) {
                 return res.status(200).json({ error: "Check your amount again" })
             }
             else {
@@ -85,7 +79,7 @@ const withdrawMoney = (req, res) => {
                 return res.status(200).json("Success")
             }
         }
-        return res.status(200).json({ error: "Check your amount again- cannot be a negative number or proide value" })
+        return res.json({ error: checkAmount(amount).error })
     }
     return res.status(200).json({ error: "Sorry, user not found " })
 };
@@ -96,11 +90,11 @@ const transferMoney = (req, res) => {
     const fromUser = users.find((user) => user.id === parseInt(fromUserId));
     console.log(fromUser);
     const toUser = users.find((user) => user.id === parseInt(toUserId));
-    if (!fromUser || !toUser || amount < 0) {
+    if (!fromUser || !toUser || amount < 0 || fromUserId === toUserId) {
         return res.status(200).json({ error: 'Sorry, check your users or amount again' })
     }
-    else if (fromUser.cash + fromUser.credit < amount) {
-        return res.status(200).json({ error: 'your cash & credit isnt enough for this transfer' })
+    else if (!checkWithdraw(fromUserId, amount)) {
+        return res.status(200).json({ error: 'your cash & credit arent enough for this transfer' })
     }
     else {
         fromUser.cash -= amount;
@@ -111,6 +105,23 @@ const transferMoney = (req, res) => {
     }
 };
 
+const checkifUserExist = (id) => {
+    const user = users.find((user) => user.id === parseInt(id));
+    if (user) return true;
+    return false;
+}
+
+const checkAmount = (amount)=>{
+    if(!amount) return { error: "please put amount" };
+    else if (amount < 0) return{ error: "amount should be a positive number" };
+    return { error: false };
+}
+
+const checkWithdraw = (id, amount) =>{
+    const user = users.find((user) => user.id === parseInt(id));
+    if (user.cash + user.credit < amount) return false;
+    return true;
+}
 
 
 module.exports = {
